feat: prepare jumpstart transaction in send amount screen
